Fix iOS autoscroll issue
On iOS chrome and iOS Safari, when you clicked on the ace editor, the page would mysteriously scroll to the top. I couldn't seem to find a good way to stop this from happening (catching the focus event and doing "e.preventDefault()" didn't seem to help). So my solution is to just scroll back down to the editor right away, which creates a slightly confusing flickering effect, but it's the best I've got at present. Better solutions welcomed.