Fix "copy successful" toasts not being visible on iOS chrome and iOS safari
As long as focus was on the curl_cmd textarea, toastr notifications would not show up in the viewport; rather they showed up at the top of the document. Automatically causing the textarea to lose focus after copying makes it so that the toastr messages show up in the top-right of the viewport like they normally do. Plus it's better to cause that field to lose focus automatically anyway, so this is good.